app.py
import streamlit as st
from Bio import SeqIO
from Bio.Seq import Seq
from Bio.SeqRecord import SeqRecord
import nglview as nv
import tempfile

def predict_structure(sequence):
    with tempfile.NamedTemporaryFile(delete=False, suffix=".fasta") as temp_fasta:
        record = SeqRecord(Seq(sequence), id="protein", description="Predicted protein structure")
        SeqIO.write(record, temp_fasta.name, "fasta")
        temp_fasta.close()
        predicted_pdb_path = temp_fasta.name.replace(".fasta", ".pdb")
        with open(predicted_pdb_path, "w") as pdb_file:
            pdb_file.write("PLACEHOLDER PDB CONTENT")
        return predicted_pdb_path

st.title("Protein Structure Prediction")

st.markdown("""
This app allows you to input an amino acid sequence and predicts the protein structure.
""")

sequence_input = st.text_area("Enter Amino Acid Sequence", height=200)

if st.button("Predict Structure"):
    if sequence_input:
        with st.spinner('Predicting protein structure...'):
            pdb_file = predict_structure(sequence_input)
            st.success("Structure prediction complete!")
            view = nv.show_file(pdb_file)
            view.render_image()
            st.pyplot(view)
        with open(pdb_file, "rb") as file:
            st.download_button(label="Download PDB File", data=file, file_name="predicted_structure.pdb", mime="chemical/x-pdb")
    else:
        st.error("Please input an amino acid sequence.")
